% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_slash_n.R
\name{add_slash_n}
\alias{add_slash_n}
\title{Adding Slash_n inside Strings}
\usage{
add_slash_n(x, delete_space = TRUE, vertical_line = TRUE)
}
\arguments{
\item{x}{a character vector}

\item{delete_space}{whether to delete spaces.
Default is TRUE.}

\item{vertical_line}{whether to 
change \code{-} into \code{|}. 
Default is TRUE.}
}
\description{
This function simply adds change-line
signs inside strings, so 
that they can be put vertically as the texts 
of x-axis.
}
\examples{
lab=add_slash_n(c("a b-c", "d - ef ", "n"))
}
