% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_click_color.R
\name{get_click_color}
\alias{get_click_color}
\title{Obtaining the Colors of Positions Clicked}
\usage{
get_click_color(x)
}
\arguments{
\item{x}{a raster object, or an image loaded by
\code{magick::image_read} or the filename of 
that image.}
}
\description{
The function draws an image and asks the user 
to click on the positions whose colors the user 
wants to know. NOTE: after clicking, you must 
press Esc button to continue. The result is a vector
of colors in hex mode.
}
