% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_locator.R
\name{image_locator}
\alias{image_locator}
\title{Get the Width and Height of the
Mouse Clicked Points}
\usage{
image_locator(x, rectangle = FALSE)
}
\arguments{
\item{x}{a raster object, or an image loaded by
\code{magick::image_read} or the filename of 
that image.}

\item{rectangle}{if it is FALSE (default), 
the result list 
contains the width and height values. 
If it is TRUE, only the left, right, top, bottom values 
of the rectangle 
designated by your clicking are returned.}
}
\description{
This function simply gets the width and 
height values of the points on which you click.
The result is a list of two vectors, the first 
vector is for width, the second for height.
}
