% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_transparent_inverse.R
\name{image_transparent_inverse}
\alias{image_transparent_inverse}
\title{Keep Certain Colors Unchanged
and Make Others Transparent}
\usage{
image_transparent_inverse(x, color, fuzz = 0)
}
\arguments{
\item{x}{a magick image.}

\item{color}{one or more colors you want 
want to keep unchanged.}

\item{fuzz}{color tolerance between 0 and 100.
Its length must be 1 or the same as \code{color}
 (which 
means you can have different fuzz values for 
different colors).
Suppose your color is white. If fuzz=0, then only 
white will be kept unchanged; if fuzz=10, colors 
similar to white will also kept unchanged.}
}
\description{
This function is an inverse version of 
\code{magick::image_transparent}.
While the latter makes certain colors 
transparent, the former keeps them 
unchanged and make others transparent.
}
