% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showcolor.R
\name{showcolor}
\alias{showcolor}
\title{Show a Color Palette}
\usage{
showcolor(x, label_size = 15, ...)
}
\arguments{
\item{x}{a character vector of colors.}

\item{label_size}{size of text on 
x-axis to show color names.}

\item{...}{other arguments passed to 
\code{geom_bar}.}
}
\description{
Simple function to show colors. NOTE: 
do not add \code{coord_flip()}.
}
\examples{
# A palette used by David Hockney
co=c("#833822", "#C03800", "#D3454C", 
	"#DC6A30", "#F29856", "#FEEF70", 
	"#A5D56D", "#16D670", "#00932F", 
	"#03592E", "#04B7B0", "#007BA9", 
	"#EC46BF", "#6A2C8F"
)
showcolor(co, label_size=10)
}
