% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_crop_click.R
\name{image_crop_click}
\alias{image_crop_click}
\title{Cuts out a Subregion of an Image by Mouse Click}
\usage{
image_crop_click(x, only_value = FALSE)
}
\arguments{
\item{x}{an image read into R by 
\code{magick::image_read} or an image  
modified by functions in the magick 
package.}

\item{only_value}{the default is FALSE, 
which will return the subregion. If you set it to 
TRUE, the result is only four values with 
the order: left, right, top, bottom.}
}
\description{
This function is a wrapper of 
\code{magick::image_crop}. While the 
latter asks you to set a \code{geometry} 
parameter, this function enables you 
to set the four sides of a subregion only 
by click the mouse. You must click at 
least 2 times (that is, click on 2 different
points to define a rectangle). After clicking, 
please press Esc on your keyboard.
}
